Extract addFile out of the add action for reuse
